+++ src/plugin/cursus/Resources/modules/course/components/sessions.jsx
@@ -13,7 +13,7 @@ import {Course as CourseTypes} from '#/plugin/cursus/prop-types'
 import {route} from '#/plugin/cursus/routing'
 import {MODAL_SESSION_FORM} from '#/plugin/cursus/session/modals/parameters'
 import {SessionList} from '#/plugin/cursus/session/components/list'
-import {selectors} from '#/plugin/cursus/tools/trainings/catalog/store/selectors'
+import {selectors} from '#/plugin/cursus/course/store'
 import {getInfo, isRegistered, isFull, canSelfRegister} from '#/plugin/cursus/utils'
 import {MODAL_COURSE_REGISTRATION} from '#/plugin/cursus/course/modals/registration'
 import {withRouter} from 'react-router-dom'
