+++ src/plugin/cursus/Resources/modules/course/containers/events.jsx
@@ -2,7 +2,7 @@ import {connect} from 'react-redux'
 
 import {actions as listActions} from '#/main/app/content/list/store'
 
-import {selectors} from '#/plugin/cursus/tools/trainings/catalog/store'
+import {selectors} from '#/plugin/cursus/course/store'
 import {CourseEvents as CourseEventsComponent} from '#/plugin/cursus/course/components/events'
 
 const CourseEvents = connect(
