+++ src/plugin/cursus/Resources/modules/course/containers/participants.jsx
@@ -1,7 +1,7 @@
 import {connect} from 'react-redux'
 
 import {constants} from '#/plugin/cursus/constants'
-import {actions, selectors} from '#/plugin/cursus/tools/trainings/catalog/store'
+import {actions, selectors} from '#/plugin/cursus/course/store'
 import {CourseParticipants as CourseParticipantsComponent} from '#/plugin/cursus/course/components/participants'
 
 const CourseParticipants = connect(
