+++ src/plugin/cursus/Resources/modules/course/store/index.js
@@ -1,0 +1,9 @@
+import {actions} from '#/plugin/cursus/course/store/actions'
+import {reducer} from '#/plugin/cursus/course/store/reducer'
+import {selectors} from '#/plugin/cursus/course/store/selectors'
+
+export {
+  actions,
+  reducer,
+  selectors
+}
