+++ src/plugin/cursus/Resources/modules/session/containers/groups.jsx
@@ -1,6 +1,6 @@
 import {connect} from 'react-redux'
 
-import {actions} from '#/plugin/cursus/tools/trainings/catalog/store'
+import {actions} from '#/plugin/cursus/course/store'
 import {SessionGroups as SessionGroupsComponent} from '#/plugin/cursus/session/components/groups'
 
 const SessionGroups = connect(
