+++ src/plugin/cursus/Resources/modules/tools/events/containers/tool.jsx
@@ -1,18 +1,23 @@
 import {connect} from 'react-redux'
-import {withReducer} from '#/main/app/store/reducer'
 
 import {hasPermission} from '#/main/app/security'
-import {actions as listActions} from '#/main/app/content/list/store'
+import {withReducer} from '#/main/app/store/reducer'
+
+import {selectors} from '#/plugin/cursus/tools/events/store'
 import {selectors as toolSelectors} from '#/main/core/tool/store'
 
-import {actions as eventActions, reducer as eventReducer, selectors as eventSelectors} from '#/plugin/cursus/event/store'
+import {actions as listActions} from '#/main/app/content/list/store'
+import {actions as courseActions} from '#/plugin/cursus/course/store'
+
 import {EventsTool as EventsToolComponent} from '#/plugin/cursus/tools/events/components/tool'
-import {selectors} from '#/plugin/cursus/tools/events/store'
+import {actions as eventActions, reducer as eventReducer, selectors as eventSelectors} from '#/plugin/cursus/event/store'
+
 
 const EventsTool = withReducer(eventSelectors.STORE_NAME, eventReducer)( // not the best place to do it
   connect(
     (state) => ({
       path: toolSelectors.path(state),
+      course: selectors.course(state),
       currentContext: toolSelectors.context(state),
       canEdit: hasPermission('edit', toolSelectors.toolData(state)),
       canRegister: hasPermission('register', toolSelectors.toolData(state))
@@ -20,6 +25,9 @@ const EventsTool = withReducer(eventSelectors.STORE_NAME, eventReducer)( // not 
     (dispatch) => ({
       open(id) {
         dispatch(eventActions.open(id))
+      },
+      openCourse(slug) {
+        dispatch(courseActions.open(slug))
       },
       invalidateList() {
         dispatch(listActions.invalidateData(selectors.LIST_NAME))
