+++ src/plugin/cursus/Resources/translations/tools.en.json
@@ -1,5 +1,5 @@
 {
-    "training_events": "Training events",
+    "training_events": "Training",
     "trainings": "Trainings",
     "presence": "Presence"
 }
