+++ src/plugin/cursus/Resources/translations/tools.fr.json
@@ -1,5 +1,5 @@
 {
-    "training_events": "Séances de formation",
+    "training_events": "Formation",
     "trainings": "Formations",
     "presence": "Présence"
 }
